Change Layout component
