+++ src/authorizationStatus.ts
@@ -1,0 +1,2 @@
+import { AuthorizationStatus } from './components/const/const';
+export const getAuthorizationStatus = () => AuthorizationStatus.NoAuth;
