+++ src/components/const/const.tsx
@@ -10,3 +10,4 @@ export enum AuthorizationStatus {
   NoAuth = 'NO_AUTH',
   Unknown = 'UNKNOWN',
 }
+
