+++ src/components/const/index.ts
@@ -1,2 +1,3 @@
 import { AppRoute } from './const';
 export default AppRoute;
+
